pyproject.toml
[project]
name = "image-similarity-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clip",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "pathlib>=1.0.1",
    "pillow>=11.2.1",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.9.1",
    "pyright>=1.1.400",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "starlette>=0.46.2",
    "torch>=2.7.0",
    "uvicorn>=0.34.2",
]

[tool.uv]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }
